Refactor: move common token substitutions to query-utils
