+++ lib/cartodb/backends/layer-stats/mapnik-layer-stats.js
@@ -1,27 +1,5 @@
-var queryUtils = require('../../utils/query-utils');
+const queryUtils = require('../../utils/query-utils');
 const AggregationMapConfig = require('../../models/aggregation/aggregation-mapconfig');
-var SubstitutionTokens = require('../../utils/substitution-tokens');
-
-// Instantiate a query with tokens for a given zoom level
-function queryForZoom(sql, zoom, singleTile=false) {
-    const tileRes = 256;
-    const wmSize = 6378137.0*2*Math.PI;
-    const nTiles = Math.pow(2, zoom);
-    const tileSize = wmSize / nTiles;
-    const resolution = tileSize / tileRes;
-    const scaleDenominator = resolution / 0.00028;
-    const x0 = -wmSize/2, y0 = -wmSize/2;
-    let bbox = `ST_MakeEnvelope(${x0}, ${y0}, ${x0+wmSize}, ${y0+wmSize})`;
-    if (singleTile) {
-        bbox = `ST_MakeEnvelope(${x0}, ${y0}, ${x0 + tileSize}, ${y0 + tileSize})`;
-    }
-    return SubstitutionTokens.replace(sql, {
-        bbox: bbox,
-        scale_denominator: scaleDenominator,
-        pixel_width: resolution,
-        pixel_height: resolution
-    });
-}
 
 function MapnikLayerStats () {
     this._types = {
@@ -52,7 +30,7 @@ function _getSQL(ctx, query, type='pre', zoom=0) {
     else {
         sql = ctx.aggrQuery;
     }
-    sql = queryForZoom(sql, zoom || 0);
+    sql = queryUtils.subsituteTokensForZoom(sql, zoom || 0);
     return query(sql);
 }
