+++ lib/cartodb/backends/overviews-metadata.js
@@ -1,4 +1,4 @@
-var SubstitutionTokens = require('../utils/substitution-tokens');
+const queryUtils = require('../utils/query-utils');
 
 function OverviewsMetadataBackend(pgQueryRunner) {
     this.pgQueryRunner = pgQueryRunner;
@@ -6,20 +6,15 @@ function OverviewsMetadataBackend(pgQueryRunner) {
 
 module.exports = OverviewsMetadataBackend;
 
-function prepareSql(sql) {
-    return sql && SubstitutionTokens.replace(sql, {
-        bbox: 'ST_MakeEnvelope(0,0,0,0)',
-        scale_denominator: '0',
-        pixel_width: '1',
-        pixel_height: '1'
-    });
-}
-
 OverviewsMetadataBackend.prototype.getOverviewsMetadata = function (username, sql, callback) {
     // FIXME: Currently using internal function _cdb_schema_name
     //        CDB_Overviews should provide the schema information directly.
-    var query = 'SELECT *, _cdb_schema_name(base_table)' +
-                '  FROM CDB_Overviews(CDB_QueryTablesText($windshaft$' + prepareSql(sql) + '$windshaft$))';
+    const query = `
+      SELECT *, _cdb_schema_name(base_table)
+        FROM CDB_Overviews(
+            CDB_QueryTablesText($windshaft$${queryUtils.substituteDummyTokens(sql)}$windshaft$)
+        );
+    `;
     this.pgQueryRunner.run(username, query, function handleOverviewsRows(err, rows) {
         if (err){
             callback(err);
