+++ lib/cartodb/models/aggregation/aggregation-mapconfig.js
@@ -8,18 +8,9 @@ const {
     createAggregationFiltersValidator
 } = aggregationValidator;
 
-const SubstitutionTokens = require('../../utils/substitution-tokens');
+const queryUtils = require('../../utils/query-utils');
 
 const removeDuplicates = arr => [...new Set(arr)];
-
-function prepareSql(sql) {
-    return sql && SubstitutionTokens.replace(sql, {
-        bbox: 'ST_MakeEnvelope(0,0,0,0)',
-        scale_denominator: '0',
-        pixel_width: '1',
-        pixel_height: '1'
-    });
-}
 
 module.exports = class AggregationMapConfig extends MapConfig {
     static get AGGREGATIONS () {
@@ -205,7 +196,7 @@ module.exports = class AggregationMapConfig extends MapConfig {
             }
 
             const sql = limitedQuery({
-                query: prepareSql(layer.options.sql)
+                query: queryUtils.substituteDummyTokens(layer.options.sql)
             });
 
             connection.query(sql, (err, result) => {
