+++ lib/cartodb/utils/query-utils.js
@@ -1,3 +1,5 @@
+const SubstitutionTokens = require('./substitution-tokens');
+
 function prepareQuery(sql) {
     var affectedTableRegexCache = {
         bbox: /!bbox!/g,
@@ -176,5 +178,37 @@ function queryPromise(dbConnection, query) {
     });
 }
 
+function substituteDummyTokens(sql) {
+    return sql && SubstitutionTokens.replace(sql, {
+        bbox: 'ST_MakeEnvelope(0,0,0,0)',
+        scale_denominator: '0',
+        pixel_width: '1',
+        pixel_height: '1'
+    });
+}
+
+function subsituteTokensForZoom(sql, zoom, singleTile=false) {
+    const tileRes = 256;
+    const wmSize = 6378137.0*2*Math.PI;
+    const nTiles = Math.pow(2, zoom);
+    const tileSize = wmSize / nTiles;
+    const resolution = tileSize / tileRes;
+    const scaleDenominator = resolution / 0.00028;
+    const x0 = -wmSize/2, y0 = -wmSize/2;
+    let bbox = `ST_MakeEnvelope(${x0}, ${y0}, ${x0+wmSize}, ${y0+wmSize})`;
+    if (singleTile) {
+        bbox = `ST_MakeEnvelope(${x0}, ${y0}, ${x0 + tileSize}, ${y0 + tileSize})`;
+    }
+    return SubstitutionTokens.replace(sql, {
+        bbox: bbox,
+        scale_denominator: scaleDenominator,
+        pixel_width: resolution,
+        pixel_height: resolution
+    });
+}
+
 module.exports.queryPromise = queryPromise;
 module.exports.getQueryLimited = getQueryLimited;
+module.exports.substituteDummyTokens = substituteDummyTokens;
+module.exports.subsituteTokensForZoom = subsituteTokensForZoom;
+
